ResendingControl: Set value before setting instance
When we set the instance before the value, there is a chance that the
spawned thread has it's first run when the instance is already set,
but the value is still at `None`. For most settings, this should not
be an issue. But by setting the value before the instance, we make
sure that there is no rogue early first call in the spawned thread.
And the backend setting only gets the intended values (and not
accidentally `None`).