BooleanStandardizedControl: Stop discretizing set value
Mapping the backend value to the reported value had allowed to detect
whether or not True/False got sent te the backend, but it was
confusing when trying to link the value to motors, as in:

```
for i in range(100):
  car.light = i # <-- this is a boolan value
  car.motor = car.light
```

As the car would not slowly increase, but jump to full speed at
`i == 50`. So we now report the standardized set value, which makes
the example work as expected.