bus/bluez: Switch scanner to passive scans
Passive scans are much quicker to report data updates, and they are
good enough for us. The quicker reporting means more work in the
filters, which even small boards can handle. But the quicker updates
allow to respond more quickly when hardware remote controls change
their output. So we can react to changes in milliseconds instead of
previously within 10 seconds, which was far too unresponsive. So we
switch to passive scans.